Add withSocket function to API.
Useful as a resource wrapper such that client code does not have to deal
with manually closing sockets. Also rename "with" to "withContext" to align
better with "withSocket". The internal withSocket function is renamed to
"onSocket" to avoid name clashes.

Change test examples to use "withSocket" and in queue.hs switch to using
forkIO instead of forkOS as with zeromq 2.1 this is safe.

Finally fix the GC finalizer for sockets to close the socket iff status is
False, which means the socket has not been closed by client code.