Rework poll.
To enable identifying sockets passed into poll, the result list is now
always of the same length, such that sockets can be identified by list
index.

PollEvent has a new constructor "None" to denote absence of events. In
case of timeout, this is put into Poll to distinguish from those sockets
for which an event occured.