Remove support for polling.
This should not be necessary as 'send' and 'receive' are internally
non-blocking and use GHC's I/O manager to block calling threads when
send or receive would yield EAGAIN. This combined with
GHC's scalable threading model should relieve client code from the
burden to do it's own polling. For timeouts there is
'System.Timeout.timeout'.