Fix socket finaliser.
Previously he finaliser was added to the Socket and it seems that with
GHC 7.4.1 that box was optimised away, causing the finaliser to run
prematurely. Attaching the finaliser to a Ptr value is as risky so with
this change the finaliser is added to the internal '_sockLive' IORef,
assuming that the lifetime of this value is the same as that of _socket.