Block all signals during zmq_term.
Contrary to zmq_term(2) which states that it is safe to restart zmq_term
if interrupted by a signal, this causes an abort when used with
OpenPGM. pgm_shutdown is reading and modifying static data which makes
it return false on second invocation which in turn makes zmq_term fail with
"Assertion failed: false (zmq.cpp:185)".

To work around this, all signals will be blocked before calling zmq_term.