Improve zmq_term signal masking.
When masking signals around the call to zmq_term, it is desirable for
some signals not to be blocked, namely termination signals such as
SIGTERM.

Also, if signal handlers are set to SIG_IGN to mask the signals during
the execution of zmq_term, it is desirable that they be restored at its
completion.  While in many applications the call to terminate the ZeroMQ
context will occur at application's end, there will be some which do
other computations after the context is released.  Therefore, this
commit restores the signal handlers after zmq_term returns.