Make zmq_device interruptible.
It is sometimes desirable to run a zmq_device call in a thread other
than the main thread.  To cleanly unwind the program, the main thread
must tell the device thread to exit thereby closing its sockets.
Unfortunately, it is not possible to interrupt a "safe" FFI call with an
asynchronous exception.  Therefore, we have change the foreign
declaration for zmq_device from "safe" to "interruptible".